refactor: PokemonCard 컴포넌트 삭제
