--- src/pages/MainList/ui/PokemonCard.module.scss
@@ -1,31 +1,0 @@
-@use "styles/color.module" as color;
-
-.card-layout {
-  width: 200px;
-  height: 210px;
-
-  padding: 8px 8px 2px;
-
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  align-items: center;
-
-  border: 1px solid map-get(color.$background, border);
-  border-radius: 12px;
-  background: radial-gradient(
-    circle,
-    map-get(color.$background, header),
-    map-get(color.$background, background-primary)
-  );
-  color: map-get(color.$common, white);
-
-  img {
-    width: 160px;
-    height: 160px;
-    visibility: hidden;
-  }
-  p {
-    height: fit-content;
-  }
-}
